refactor: extract task creation service
