+++ apps/api/src/lib/entity-versions/version-write-ownership-policy.ts
@@ -47,7 +47,7 @@ export const REVIEWED_VERSION_MUTATION_OWNERS = {
     VERSION_WRITE_CAPABILITY.SET_CURRENT_VERSION,
     VERSION_WRITE_CAPABILITY.USE_VERSION_UTILS,
   ],
-  "handlers/tasks/create.ts": [
+  "lib/tasks/create-task-entity.ts": [
     VERSION_WRITE_CAPABILITY.INSERT_VERSION_ROW,
     VERSION_WRITE_CAPABILITY.SET_CURRENT_VERSION,
   ],
